feat: streamline agent addition and prompt editing
This commit reworks the agent addition process and prompt editing
in the crew panel. It fixes issues with adding agents to the crew.
The prompt node has been removed and replaced with a prompt editor
in the crew panel.
