+++ apps/web/src/routes/app/crews/[id]/nodes/index.ts
@@ -1,10 +1,7 @@
 import Agent from './node-agent.svelte';
-import Prompt from './node-prompt.svelte';
 
 export {
 	Agent,
-	Prompt,
 	//
-	Agent as NodeAgent,
-	Prompt as NodePrompt
+	Agent as NodeAgent
 };
